+++ algorithm/Algorithm.hpp
@@ -42,10 +42,11 @@ protected:
     double hitrate;
 #endif
 
-    std::tuple<int, int> calculateInteractions(std::vector<Utility::Particle> &b0, std::vector<Utility::Particle> &b1,
-                                               std::vector<Utility::Particle> &b2, int b0Owner, int b1Owner,
-                                               int b2Owner, int b0Start, int b0NumSteps, double cutoff,
-                                               Eigen::Array3d physicalDomainSize);
+    std::tuple<uint64_t, uint64_t> calculateInteractions(std::vector<Utility::Particle> &b0,
+                                                         std::vector<Utility::Particle> &b1,
+                                                         std::vector<Utility::Particle> &b2, int b0Owner, int b1Owner,
+                                                         int b2Owner, int b0Start, int b0NumSteps, double cutoff,
+                                                         Eigen::Array3d physicalDomainSize);
 #ifdef PROFILE_3BMDA
     void calcParticleInteractions(std::vector<std::tuple<int, int, int>> &particleTripletsToCalculate,
                                   std::vector<Utility::Particle> &b0, std::vector<Utility::Particle> &b1,
@@ -61,17 +62,20 @@ public:
     virtual ~Algorithm();
 
     virtual void Init(std::shared_ptr<Simulation> simulation);
-    virtual std::tuple<int, int> SimulationStep() = 0;
+    virtual std::tuple<uint64_t, uint64_t> SimulationStep() = 0;
 
-    std::tuple<int, int> CalculateInteractions(std::vector<Utility::Particle> &b0, std::vector<Utility::Particle> &b1,
-                                               std::vector<Utility::Particle> &b2, int b0Owner, int b1Owner,
-                                               int b2Owner);
-    std::tuple<int, int> CalculateInteractions(std::vector<Utility::Particle> &b0, std::vector<Utility::Particle> &b1,
-                                               std::vector<Utility::Particle> &b2, int b0Owner, int b1Owner,
-                                               int b2Owner, int b0Start, int b0NumSteps);
-    std::tuple<int, int> CalculateInteractions(std::vector<Utility::Particle> &b0, std::vector<Utility::Particle> &b1,
-                                               std::vector<Utility::Particle> &b2, int b0Owner, int b1Owner,
-                                               int b2Owner, double cutoff, Eigen::Array3d physicalDomainSize);
+    std::tuple<uint64_t, uint64_t> CalculateInteractions(std::vector<Utility::Particle> &b0,
+                                                         std::vector<Utility::Particle> &b1,
+                                                         std::vector<Utility::Particle> &b2, int b0Owner, int b1Owner,
+                                                         int b2Owner);
+    std::tuple<uint64_t, uint64_t> CalculateInteractions(std::vector<Utility::Particle> &b0,
+                                                         std::vector<Utility::Particle> &b1,
+                                                         std::vector<Utility::Particle> &b2, int b0Owner, int b1Owner,
+                                                         int b2Owner, int b0Start, int b0NumSteps);
+    std::tuple<uint64_t, uint64_t> CalculateInteractions(std::vector<Utility::Particle> &b0,
+                                                         std::vector<Utility::Particle> &b1,
+                                                         std::vector<Utility::Particle> &b2, int b0Owner, int b1Owner,
+                                                         int b2Owner, double cutoff, Eigen::Array3d physicalDomainSize);
 
     void SumUpParticles(std::vector<Utility::Particle> &b0, std::vector<Utility::Particle> &b1,
                         std::vector<Utility::Particle> &b2);
